Solution to Clear nth Bit of a Number

+++ Clear nth Bit of a Number@@67882a731b4392d8fc1577ec/code.c
@@ -12,7 +12,7 @@ int main() {
 
     }
     if(c==1){
-        double d=a-pow(2,b);
+        int d=a-(int)pow(2,b);
         printf("%d",d);
     }
